sap_scrapper/requirements.txt
# Core
python-dotenv>=0.19.0
pydantic>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0

# Web Scraping
playwright>=1.30.0
asyncio>=3.4.3
backoff>=2.2.1

# Base de datos
motor>=3.1.1
pymongo>=4.3.3

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-playwright>=0.3.0

# Utilidades
python-json-logger>=2.0.7 